==> notebooks/modeling/01_priceMovement/01-test-data.ipynb <==
%run ./../../utils/commonImports.py
%matplotlib inline

# Load Data

db = 'D:\\Dropbox\\My work\\krypl-project\\sqlite\\ploniex-chart-data\\USDT_BTC.db'
data = load_trading_data(db, 'chart_data', from_date='2014-01-01', period='5min')

data.head()

# Timestamp diff

diffs = data['timestamp'].shift(-1) - data['timestamp']
data['timestampDiff'] = diffs

diffs.value_counts()

zero_diff = data.query('timestampDiff == 1200')
zero_diff

data.loc[91050-10:91050+10, :]
