==> notebooks/modeling/01_priceMovement/03-price-movement-analysis.ipynb <==
%run ./../../utils/commonImports.py
%run ./../../utils/tradingImports.py
%matplotlib inline

# Load Data

df_all = read_tsv('D:\\Dropbox\\My work\\krypl-project\\analysis\\return-01-within-day.tsv')
df = df_all[['date', 'maxReturn', 'tillMax', 'minReturn']].head(df_all.shape[0] - 2)

df.query('maxReturn >= 0.01').shape[0] / df.shape[0]

# Basic Stats

query = '(maxReturn > 0.01)'
df_true = df.query(query)
df_false = df.query(f'not {query}')

df_true.describe()

df_false.describe()
